Fix hard fault due to uninitialized MPU BSS section
In the MPU protected BSS section, there is a flag indicating whether static instance has been created or not.
The flag is not cleared and during a power interruption, the flag sometimes has a random value.
This caused the SafeSupervisor static object to have random incorrect values due to missing constructor call,
which leads to unintended behaviour, e.g. false positive safety checks.

The solution is to zero out MPU protected BSS section.

Change-Id: I5de341b0fd14a64b0c5b21a8b82a1ce30c6cfba8