Fix Watchdog constructor
The actual enableWatchdog call in the ctor is wrongly using the
clockSpeed as the interruptActive argument.

Instead, we need to add the interruptActive flag in the middle. Using
the default (false) as in the enableWatchdog default value. Risk is low,
since the Watchdog class is currently not instantiated. This way,
the intention of the ctor and enableWatchdog() is synchronized regarding
the default values.