Remove redundant help for lifecycle control command (lc)
In the help menu, the "lc" command info is printed twice: Once triggered
from the SafetySystem and once from SysAdminSystem. This change removes
the one in SafetySystem.

The list of commands in _commands is supposed to make entries unique,
by using an etl::intrusive_forward_list. However, since the Systems are
instantiating the control commands individually, that does not make them
unique really.