Avoid initializing global vars of libsafe*.a before mpu is enabled
linker script already asserts that .data section of libsafe*.a is empty.
However, .rodata might be merged into .text as a compiler optimization.
So those .rodata vars are initialized before mpu being enabled.
This commit extends the assertion to check .rodata . And move the global
vars initialization to MemoryProtection::init() .

Signed-off-by: Chiang, Yi <Yi.Chiang@bmw.de>