Switch to pre-built Docker Image
With the introduction of a pre-built docker image this commits prepares
a docker-compose file for easy local development. Also the CI workflows
now use docker compose to run the different commands. This has the
advantage, that it is easy for a developer to reproduce and we have a
central place to define the image URL to use.