Refactor Toolchain Files and fix Sysroot
While integrating the pre-build docker we noticed that we currently have
quite a tricky selection of compilers, especially for Clang. We have
both the x86 and the ARM Clang in the path and it just selects "by
accident" the one which is first in path. While this selection is
deterministic it can lead to quick confusions. Note, that form GCC this
is not visible since GCC for ARM is named differently, i.e.
arm-none-eabi-xxx.

To avoid confusion, we refactored to only have the x86 compilers are in
the path and other compilers need to be specified by their full path.
This would also allow us to have different compiler versions installed
in parallel.

While changing this we noticed some issues in the toolchain files we
already have and try to clean this up.

- Only specify CMAKE_C_COMPILER in the toolchain to its full path in
  case the CC environment variable is not set. Otherwise, rely on CMake
  to properly determine the compiler.
- Rely on CMake to determine all the other tools like the C++ compiler,
  ASM compiler, ... instead of manually specifying a ton of tools.
- Remove the explictily set sysroot which is not necessary, since Clang
  automatically deduces the right one based on the other flags like
  -mcpu, -mfloat-abi or -mfpu. The currently specified sysroot was even
  wrong, because it was pointing to the soft FPU setting but the
  -mfloat-abi is set to hard.