Add Image Digest to CMake Cache
From now on we will take the digest of the built docker image as part of
the cmake cache key. Otherwise, changes to the docker image would not
result in a re-build but in taking the cached artifacts. In case the
change in the Dockerfile would actually break the build we would not
see this because nothing is executed.