Make sure launchCommandFromRAM isn't inlined
When flashing the freertos version of referenceApp on top of the threadX
version (or vice-versa), we noticed a hard fault on first startup.
This was triggered by an eeprom error, caused by accessing the code
flash while an eeprom configuration command was running.

Looking at the dissasembly, we noticed launchCommandFromRAM had been
inlined and was thus running in launchCommand, from ROM.

Change-Id: Idb3283e11b9164fef0cffa59478c77c387ac06da